refactor: Move ggshield.cmd.main to ggschield.__main__
This will simplify our import linter rules and allow to debug ggshield
more easily using a debugger.

refs #521
