+++ tests/unit/cmd/auth/test_logout.py
@@ -4,7 +4,7 @@ from unittest.mock import Mock
 import pytest
 from requests.exceptions import ConnectionError
 
-from ggshield.cmd.main import cli
+from ggshield.__main__ import cli
 from ggshield.core.config import Config
 from ggshield.core.constants import DEFAULT_INSTANCE_URL
 from ggshield.core.errors import ExitCode
