+++ tests/unit/cmd/hmsl/test_api_status.py
@@ -3,7 +3,7 @@ from pathlib import Path
 import pytest
 from click.testing import CliRunner
 
-from ggshield.cmd.main import cli
+from ggshield.__main__ import cli
 from tests.unit.conftest import assert_invoke_ok, my_vcr
