+++ tests/unit/cmd/iac/test_scan_diff.py
@@ -2,7 +2,7 @@ from pathlib import Path
 
 from click.testing import CliRunner
 
-from ggshield.cmd.main import cli
+from ggshield.__main__ import cli
 from ggshield.core.errors import ExitCode
 from ggshield.verticals.secret.repo import cd
 from tests.conftest import _IAC_SINGLE_VULNERABILITY
