+++ tests/unit/cmd/sca/test_precommit.py
@@ -3,7 +3,7 @@ from unittest.mock import patch
 
 from click.testing import CliRunner
 
-from ggshield.cmd.main import cli
+from ggshield.__main__ import cli
 from ggshield.core.errors import ExitCode
 from ggshield.verticals.sca.sca_scan_models import (
     ComputeSCAFilesResult,
