+++ tests/unit/cmd/scan/test_path.py
@@ -5,7 +5,7 @@ from pathlib import Path
 import pytest
 from click.testing import CliRunner
 
-from ggshield.cmd.main import cli
+from ggshield.__main__ import cli
 from ggshield.core.errors import ExitCode
 from tests.unit.conftest import (
     _ONE_LINE_AND_MULTILINE_PATCH,
