+++ tests/unit/cmd/scan/test_prepush.py
@@ -4,7 +4,7 @@ from unittest.mock import ANY, Mock, patch
 import pytest
 from click.testing import CliRunner
 
-from ggshield.cmd.main import cli
+from ggshield.__main__ import cli
 from ggshield.core.errors import ExitCode
 from ggshield.core.filter import init_exclusion_regexes
 from ggshield.core.scan import ScanContext, ScanMode
