+++ tests/unit/cmd/scan/test_repo.py
@@ -1,4 +1,4 @@
-from ggshield.cmd.main import cli
+from ggshield.__main__ import cli
 from ggshield.core.errors import ExitCode
 from tests.unit.conftest import assert_invoke_exited_with
